pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "Personal knowledge system for data leadership content"
requires-python = ">=3.11"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",

    # Azure
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
    "azure-functions>=1.17.0",

    # Database (Azure SQL)
    "pyodbc>=5.0.0",

    # PDF Processing
    "pymupdf>=1.23.0",

    # Claude API
    "anthropic>=0.18.0",

    # App
    "streamlit>=1.30.0",

    # Utilities
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
